improve delete groups

+++ src/main/kotlin/com/cocktailfellow/group/DeleteGroup.kt
@@ -5,8 +5,11 @@ import com.cocktailfellow.AbstractRequestHandler
 import com.cocktailfellow.ApiGatewayResponse
 import com.cocktailfellow.common.HttpStatusCode
 import com.cocktailfellow.common.JsonConfig
+import com.cocktailfellow.common.ValidationException
+import com.cocktailfellow.common.database.UserGroupLinkRepository
 import com.cocktailfellow.group.database.GroupRepository
 import com.cocktailfellow.token.TokenManagement
+import com.cocktailfellow.token.TokenManagementData
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 
@@ -20,8 +23,11 @@ class DeleteGroup : AbstractRequestHandler() {
     val request = JsonConfig.instance.decodeFromString<DeleteGroupRequest>(body)
     val groupId = request.groupId
 
-    TokenManagement.validateTokenOnly(authorization)
+    val tokenManagementData: TokenManagementData = TokenManagement.validateTokenAndGetData(authorization)
 
+    if (!UserGroupLinkRepository.isMemberOfGroup(tokenManagementData.username, groupId)) {
+      throw ValidationException("User is not allowed to delete the group.") // todo: refactor
+    }
     GroupRepository.deleteGroup(groupId)
 
     return generateResponse(HttpStatusCode.OK.code)
